requirements.txt
dill==0.3.7
fastapi==0.109.2
langchain==0.1.6
langsmith==0.0.87
magic-filter==1.0.12
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.7
numpy==1.26.0
openai==0.28.1
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.6.3
pyparsing==3.1.1
python-dotenv==1.0.0
redis==5.0.1
SQLAlchemy==2.0.21
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.24.0.post1
yarl==1.9.2
